Logger now is initialized independently
